+++ plugins/node/instrumentation-socket.io/package.json
@@ -59,7 +59,6 @@
     "typescript": "4.3.5"
   },
   "dependencies": {
-    "@opentelemetry/core": "^1.0.0",
     "@opentelemetry/instrumentation": "^0.32.0",
     "@opentelemetry/semantic-conventions": "^1.0.0"
   },
