+++ plugins/node/instrumentation-socket.io/tsconfig.json
@@ -3,7 +3,6 @@
   "compilerOptions": {
     "rootDir": ".",
     "outDir": "build",
-    "skipLibCheck": true
   },
   "include": [
     "src/**/*.ts",
